+++ kamehouse-shell/bin/common/functions/extended-functions.sh
@@ -3,48 +3,6 @@ if (( $EUID == 0 )); then
   IS_ROOT_USER=true
 fi
 
-# Subsystem root prefix for mounted drives. Use this as a prefix to all
-# absolute paths I define in the script.
-ROOT_PREFIX="/mnt"
-
-# Update this function both in common-functions.sh and path.sh
-export IS_LINUX_HOST=false
-
-setIsLinuxHost() {
-  export IS_LINUX_HOST=false
-  local UNAME_S=`uname -s`
-  local UNAME_R=`uname -r`
-  if [ "${UNAME_S}" != "Linux" ]; then
-    # Using Git Bash
-    export IS_LINUX_HOST=false
-  else 
-    if [[ ${UNAME_R} == *"Microsoft"* ]]; then
-      # Using Ubuntu for Windows 10 (deprecated. don't use that anymore, use an ubuntu vm)
-      export IS_LINUX_HOST=false
-    else
-      # Using Linux
-      export IS_LINUX_HOST=true
-    fi
-  fi
-}
-setIsLinuxHost
-
-# Check if I'm using Ubuntu for windows, Git Bash or any other bash implementation.
-# Default is Ubuntu for windows. Set root prefix for mounted drives based on the subsystem.
-# In Git Bash drives are mounted /c /d so root prefix is empty.
-# In Ubuntu for windows drives are mounted in /mnt/c /mnt/d so root prefix is /mnt
-setRootPrefix() {
-  # Ubuntu for windows
-  ROOT_PREFIX="/mnt"
-  local MSYSTEM_MINGW="${MSYSTEM}"
-  MSYSTEM_MINGW=`echo ${MSYSTEM_MINGW:0:5}`
-  if [ "${MSYSTEM_MINGW}" == "MINGW" ]; then 
-    # Git Bash
-    ROOT_PREFIX=""
-  fi
-}
-setRootPrefix
-
 urlencode() {
     # urlencode <string>
     old_lc_collate=$LC_COLLATE