+++ kamehouse-shell/bin/is-linux-host.sh
@@ -7,10 +7,13 @@ if [ "$?" != "0" ]; then
 	exit 99
 fi
 
+LOG=DISABLED
+LOG_PROCESS_TO_FILE=false
+
 # Returns true if it's a linux host, false if it isn't.
 # It can also be infered by the return value. 
 # Returns 0 for true, 1 for false.
-main() {
+mainProcess() {
 	echo ${IS_LINUX_HOST}
   if ${IS_LINUX_HOST}; then
     exit ${EXIT_SUCCESS}