+++ kamehouse-shell/bin/kamehouse/get-kamehouse-secret.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-# Disable logs
-LOG=ERROR
-SKIP_LOG_START_FINISH=true
-
 # Import kamehouse functions
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then
@@ -11,8 +7,9 @@ if [ "$?" != "0" ]; then
   exit 99
 fi
 
-LOAD_KAMEHOUSE_SECRETS=true
+LOG=DISABLED
 LOG_PROCESS_TO_FILE=false
+LOAD_KAMEHOUSE_SECRETS=true
 KAMEHOUSE_SECRET_NAME=""
 
 mainProcess() {