+++ kamehouse-shell/bin/kamehouse/tomcat-stop.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-# Disable logs
-LOG=ERROR
-SKIP_LOG_START_FINISH=true
-
 # Import kamehouse functions
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then
@@ -11,7 +7,9 @@ if [ "$?" != "0" ]; then
   exit 99
 fi
 
-main() {
+LOG=DISABLED
+
+mainProcess() {
   if ${IS_LINUX_HOST}; then
     ${HOME}/programs/kamehouse-shell/bin/lin/kamehouse/tomcat-stop.sh "$@"
   else